Update useProductsQuery hook
