--- packages/commerce/frontend/src/lib/api/types.ts
@@ -1,5 +1,0 @@
-export interface ErrorResponse {
-  message: string
-  success: false
-  errors?: Record<string, string>
-}
