+++ packages/commerce/frontend/src/lib/utils.ts
@@ -1,8 +1,8 @@
 import { type CxOptions, cx } from 'class-variance-authority'
 import { twMerge } from 'tailwind-merge'
 import { API_BASE_URL } from './constants.ts'
-import { ErrorResponse } from './api/types.ts'
-import { APIError } from './api/error.ts'
+import { ErrorResponse, SuccessReponse } from './types.ts'
+import { APIError } from './api-error.ts'
 
 export function cn(...inputs: CxOptions) {
   return twMerge(cx(inputs))
@@ -11,11 +11,11 @@ export function cn(...inputs: CxOptions) {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export async function fetcher<JSON = any>(
   input: string,
-  init?: RequestInit
-): Promise<JSON> {
+  options?: RequestInit
+): Promise<SuccessReponse<JSON>> {
   const response = await fetch(API_BASE_URL + input, {
     credentials: 'include',
-    ...init
+    ...options
   })
   const json = await response.json()
   if (!response.ok) {
